+++ src/main/kotlin/com/hvz/controllers/UserController.kt
@@ -1,7 +1,5 @@
 package com.hvz.controllers
 
-import com.hvz.exceptions.GameNotFoundException
-import com.hvz.exceptions.UserNotFoundException
 import com.hvz.models.User
 import com.hvz.services.game.GameService
 import com.hvz.services.user.UserService
@@ -12,7 +10,6 @@ import org.springframework.web.bind.annotation.CrossOrigin
 import org.springframework.web.bind.annotation.DeleteMapping
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.PathVariable
-import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.PutMapping
 import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RestController
@@ -29,20 +26,17 @@ class UserController(private val userService: UserService,
 ) {
 
     @DeleteMapping("api/v1/users/{id}")
-    fun deleteById(@PathVariable(name = "id") uid: String): ResponseEntity<Any> {
-
-        return try {
-            userService.findById(uid)
-            userService.deleteById(uid)
-
-            ResponseEntity.noContent().build()
-        } catch (userNotFoundException: UserNotFoundException) {
-            ResponseEntity.badRequest().build()
+    fun deleteById(@PathVariable(name = "id") uid: String): ResponseEntity<Any> =
+        when (userService.findById(uid)) {
+            null -> ResponseEntity.badRequest().build()
+            else -> {
+                userService.deleteById(uid)
+                ResponseEntity.noContent().build()
+            }
         }
-    }
 
     @PutMapping("/register")
-    fun addUser(@AuthenticationPrincipal jwt: Jwt): ResponseEntity<Any> {
+    fun registerUser(@AuthenticationPrincipal jwt: Jwt): ResponseEntity<Any> {
 
         val url = URL(jwt.getClaimAsStringList("aud")[1])
         val con = url.openConnection().apply {
@@ -61,56 +55,48 @@ class UserController(private val userService: UserService,
                 split[0] to split[1]
             }
 
-        try {
-            val foundUser = userService.getUserBySub(jwt.claims["sub"] as String)
-            val updatedUser = foundUser.copy(
-                name = idClaims["name"]!!,
-                email = idClaims["email"]!!
-            )
-
-            userService.update(updatedUser)
-
-            return ResponseEntity.ok(updatedUser.toReadDto())
-        } catch (_: UserNotFoundException) { }
-
-        val user = userService.add(
-                User(idClaims["sub"]!!.removePrefix("auth0|"),
+        return when (val foundUser = userService.getUserBySub(jwt.claims["sub"] as String)) {
+            null -> {
+                val user = userService.add(
+                    User(idClaims["sub"]!!.removePrefix("auth0|"),
                         idClaims["name"]!!,
                         idClaims["email"]!!
+                    )
                 )
-        )
 
-        val uri = URI.create("api/v1/users/${user.uid}")
+                val uri = URI.create("api/v1/users/${user.uid}")
 
-        return ResponseEntity.created(uri).body(user.toReadDto())
+                ResponseEntity.created(uri).body(user.toReadDto())
+            }
+            else -> {
+                val updatedUser = foundUser.copy(name = idClaims["name"]!!, email = idClaims["email"]!!)
+                userService.update(updatedUser)
+
+                ResponseEntity.ok(updatedUser.toReadDto())
+            }
+        }
     }
 
     @GetMapping("api/v1/users")
     fun findAll() = ResponseEntity.ok(userService.findAll().map { it.toReadDto() })
 
     @GetMapping("api/v1/users/{id}")
-    fun findById(@PathVariable(name = "id") uid: String): ResponseEntity<Any> {
-
-        return try {
-            return ResponseEntity.ok(userService.findById(uid).toReadDto())
-        } catch (userNotFoundException: UserNotFoundException) {
-            ResponseEntity.notFound().build()
+    fun findById(@PathVariable(name = "id") uid: String): ResponseEntity<Any> =
+        when (val user = userService.findById(uid)) {
+            null -> ResponseEntity.notFound().build()
+            else -> ResponseEntity.ok(user.toReadDto())
         }
-    }
 
     @GetMapping("api/v1/games/{game_id}/users")
-    fun findByGameId(@PathVariable(name = "game_id") gameId: Int): ResponseEntity<Any> {
-
-        return try {
-            val game = gameService.findById(gameId)
-
-            val users = userService.findAll().filter { user ->
-                game.players.find { it.user!!.uid == user.uid } != null
-            }
-
-            return ResponseEntity.ok(users.map { it.toReadDto() })
-        } catch (gameNotFoundException: GameNotFoundException) {
-            ResponseEntity.notFound().build()
+    fun findByGameId(@PathVariable(name = "game_id") gameId: Int): ResponseEntity<Any> =
+        when (val game = gameService.findById(gameId)) {
+            null -> ResponseEntity.notFound().build()
+            else -> ResponseEntity.ok(
+                userService.findAll().filter { user ->
+                    game.players.find { it.user!!.uid == user.uid } != null
+                }.map {
+                    it.toReadDto()
+                }
+            )
         }
-    }
 }
